refactor price to float

+++ pkg/api/exchangeApi.go
@@ -7,16 +7,16 @@ import (
 )
 
 type ExchangeApi interface {
-	GetCurrentCoinPriceForFutures(coin *domains.Coin) (int64, error)
-	GetCurrentCoinPrice(coin *domains.Coin) (int64, error)
+	GetCurrentCoinPriceForFutures(coin *domains.Coin) (float64, error)
+	GetCurrentCoinPrice(coin *domains.Coin) (float64, error)
 	GetKlines(coin *domains.Coin, interval string, limit int, fromTime time.Time) (KlinesDto, error)
 	GetKlinesFutures(coin *domains.Coin, interval string, limit int, fromTime time.Time) (KlinesDto, error)
 
-	BuyCoinByMarket(coin *domains.Coin, amount float64, price int64) (OrderResponseDto, error)
-	SellCoinByMarket(coin *domains.Coin, amount float64, price int64) (OrderResponseDto, error)
+	BuyCoinByMarket(coin *domains.Coin, amount float64, price float64) (OrderResponseDto, error)
+	SellCoinByMarket(coin *domains.Coin, amount float64, price float64) (OrderResponseDto, error)
 
-	OpenFuturesOrder(coin *domains.Coin, amount float64, price int64, futuresType futureType.FuturesType, stopLossPriceInCents int64) (OrderResponseDto, error)
-	CloseFuturesOrder(coin *domains.Coin, openedTransaction *domains.Transaction, price int64) (OrderResponseDto, error)
+	OpenFuturesOrder(coin *domains.Coin, amount float64, price float64, futuresType futureType.FuturesType, stopLossPriceInCents float64) (OrderResponseDto, error)
+	CloseFuturesOrder(coin *domains.Coin, openedTransaction *domains.Transaction, price float64) (OrderResponseDto, error)
 	IsFuturesPositionOpened(coin *domains.Coin, openedOrder *domains.Transaction) bool
 	GetCloseTradeRecord(coin *domains.Coin, openTransaction *domains.Transaction) (OrderResponseDto, error)
 	GetLastFuturesOrder(coin *domains.Coin, clientOrderId string) (OrderResponseDto, error)
@@ -30,9 +30,9 @@ type ExchangeApi interface {
 }
 
 type OrderResponseDto interface {
-	CalculateAvgPrice() int64
-	CalculateTotalCost() int64
-	CalculateCommissionInUsd() int64
+	CalculateAvgPrice() float64
+	CalculateTotalCost() float64
+	CalculateCommissionInUsd() float64
 	GetAmount() float64
 	GetCreatedAt() *time.Time
 }
@@ -47,12 +47,12 @@ type KlineDto interface {
 	GetInterval() string
 	GetStartAt() time.Time
 	GetCloseAt() time.Time
-	GetOpen() int64
-	GetHigh() int64
-	GetLow() int64
-	GetClose() int64
+	GetOpen() float64
+	GetHigh() float64
+	GetLow() float64
+	GetClose() float64
 }
 
 type WalletBalanceDto interface {
-	GetAvailableBalanceInCents() int64
+	GetAvailableBalanceInCents() float64
 }
